ibeam aborts on hook errors
ibeam was aborting when a bash script exited with non-zero. However
some hooks were not exiting properly.  writing the spawned shell's
stderr and stdout to was not happening. Added ability to collect
stdout/err from process even after exit of non-zero and provide a new
option, use_stderr. This will use stderr of ibeam if an error occurs
in the hook and it is currently aborting